requirements.txt
bottle==0.12.25
gitpython == 3.1.27
requests == 2.31.0
paste==3.7.1
icecream==2.1.3

